add permission checks for user api functions

+++ src/features/users/actions/user.ts
@@ -1,5 +1,7 @@
 "use server";
 
+import { getCurrentUser } from "@/lib/currentUser";
+import { hasPermission } from "@/permissions/general";
 import {
   deleteUser as deleteUserDb,
   updateUser as updateUserDb,
@@ -7,6 +9,14 @@ import {
 
 export async function deleteUser(id: number) {
   try {
+    const user = await getCurrentUser();
+    const canDeleteUser = hasPermission(user?.role, "deleteUser");
+    if (!canDeleteUser) {
+      return {
+        error: true,
+        message: "You do not have permission to delete users.",
+      };
+    }
     await deleteUserDb({ id });
     return { error: false, message: "Successfully deleted the user" };
   } catch (error) {
@@ -15,6 +25,16 @@ export async function deleteUser(id: number) {
 }
 
 export async function updateUser({ id }: { id: number }, data: FormData) {
+  const user = await getCurrentUser();
+  const canUpdateUser = hasPermission(user?.role, "updateUser");
+
+  if (!canUpdateUser) {
+    return {
+      error: true,
+      message: "You do not have permission to update users.",
+    };
+  }
+
   const userData = {
     firstName: data.get("firstName") as string,
     lastName: data.get("lastName") as string,
